@@ -72,10 +72,10 @@
     "@azure/communication-calling": "1.13.0-beta.4 || >=1.12.1",
     "@azure/communication-calling-effects": "1.0.0-beta.2",
     "@azure/communication-chat": "1.3.2-beta.1 || >=1.2.0",
-    "@types/react": ">=16.8.0 <18.0.0",
-    "@types/react-dom": ">=16.8.0 <18.0.0",
-    "react": ">=16.8.0 <18.0.0",
-    "react-dom": ">=16.8.0 <18.0.0"
+    "@types/react": ">=16.8.0 <19.0.0",
+    "@types/react-dom": ">=16.8.0 <19.0.0",
+    "react": ">=16.8.0 <19.0.0",
+    "react-dom": ">=16.8.0 <19.0.0"
   },
   "devDependencies": {
     "@azure/communication-calling": "1.13.0-beta.4 || >=1.12.1",
@@ -91,17 +91,16 @@
     "@microsoft/api-extractor": "~7.18.0",
     "@playwright/test": "~1.22.2",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "12.1.5",
-    "@testing-library/react-hooks": "^3.4.2",
+    "@testing-library/react": "14.0.0",
     "@types/copy-webpack-plugin": "^6.4.0",
     "@types/express": "^4.17.8",
     "@types/jest": "~26.0.22",
     "@types/json-stringify-safe": "^5.0.0",
     "@types/node": "^14.14.10",
     "@types/puppeteer": "~5.4.3",
-    "@types/react-dom": "^16.9.8",
+    "@types/react-dom": "18.2.1",
     "@types/react-linkify": "^1.0.0",
-    "@types/react": "^16.9.49",
+    "@types/react": "18.2.0",
     "@types/uuid": "^8.3.0",
     "@types/yargs": "17.0.10",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
@@ -143,10 +142,9 @@
     "prettier": "2.3.1",
     "pretty-quick": "^3.1.0",
     "puppeteer": "~10.0.0",
-    "react-dom": "16.13.1",
+    "react-dom": "18.2.0",
     "react-is": "~17.0.1",
-    "react-test-renderer": "^16.14.0",
-    "react": "~16.14.0",
+    "react": "18.2.0",
     "regenerator-runtime": "^0.13.7",
     "rimraf": "^2.6.2",
     "rollup": "~2.42.4",
