@@ -41,19 +41,17 @@
     "@internal/react-composites": "1.5.1-beta.3",
     "@types/react-linkify": "^1.0.0",
     "json-stringify-safe": "^5.0.1",
-    "react": "~16.14.0",
-    "react-dom": "16.13.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-linkify": "^1.0.0-alpha",
     "reselect": "~4.0.0"
   },
   "devDependencies": {
     "@babel/core": "~7.16.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react-hooks": "^3.4.2",
     "@types/jest": "~26.0.22",
     "@types/json-stringify-safe": "^5.0.0",
-    "@types/react-dom": "^16.9.8",
-    "@types/react": "^16.9.49",
+    "@types/react-dom": "18.2.1",
+    "@types/react": "18.2.0",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
     "@typescript-eslint/parser": "^4.12.0",
     "babel-eslint": "^10.0.0",
@@ -80,7 +78,6 @@
     "jest": "26.6.0",
     "prettier": "2.3.1",
     "pretty-quick": "^3.1.0",
-    "react-test-renderer": "^16.14.0",
     "rimraf": "^2.6.2",
     "style-loader": "~2.0.0",
     "ts-loader": "^8.0.12",
