@@ -27,15 +27,15 @@
     "@azure/communication-identity": "~1.1.0",
     "@azure/communication-react": "1.5.1-beta.3",
     "@fluentui/react": "~8.98.3",
-    "react": "~16.14.0",
-    "react-dom": "16.13.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "dotenv": "~10.0.0",
     "uuid": "^8.1.0"
   },
   "devDependencies": {
     "@babel/core": "~7.16.0",
-    "@types/react": "^16.9.49",
-    "@types/react-dom": "^16.9.8",
+    "@types/react": "18.2.0",
+    "@types/react-dom": "18.2.1",
     "@types/uuid": "^8.3.0",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
     "@typescript-eslint/parser": "^4.12.0",
