@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { initializeIconsForUITests } from '../lib/utils';
 import { HermeticApp, shouldLoadHermeticApp } from './HermeticApp';
 import { LiveApp } from './LiveApp';
@@ -11,7 +11,11 @@ import { parseQueryArgs } from './QueryArgs';
 initializeIconsForUITests();
 const queryArgs = parseQueryArgs();
 
-ReactDOM.render(
-  shouldLoadHermeticApp(queryArgs) ? <HermeticApp queryArgs={queryArgs} /> : <LiveApp queryArgs={queryArgs} />,
-  document.getElementById('root')
+const domNode = document.getElementById('root');
+if (!domNode) {
+  throw new Error('Failed to find the root element');
+}
+
+createRoot(domNode).render(
+  shouldLoadHermeticApp(queryArgs) ? <HermeticApp queryArgs={queryArgs} /> : <LiveApp queryArgs={queryArgs} />
 );
