@@ -2,16 +2,20 @@
 // Licensed under the MIT license.
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './app/App';
 import { SwitchableFluentThemeProvider } from './app/theming/SwitchableFluentThemeProvider';
 
-ReactDOM.render(
+const domNode = document.getElementById('root');
+if (!domNode) {
+  throw new Error('Failed to find the root element');
+}
+
+createRoot(domNode).render(
   <SwitchableFluentThemeProvider scopeId="SampleCallingApp">
     <div className="wrapper">
       <App />
     </div>
-  </SwitchableFluentThemeProvider>,
-  document.getElementById('root')
+  </SwitchableFluentThemeProvider>
 );
