@@ -2,16 +2,18 @@
 // Licensed under the MIT license.
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import { Examples } from './Examples';
 import { _IdentifierProvider } from '@azure/communication-react';
 
-if (document.getElementById('root') !== undefined) {
-  ReactDOM.render(
-    <div className="wrapper">
-      <Examples />
-    </div>,
-    document.getElementById('root')
-  );
+const domNode = document.getElementById('root');
+if (!domNode) {
+  throw new Error('Failed to find the root element');
 }
+
+createRoot(domNode).render(
+  <div className="wrapper">
+    <Examples />
+  </div>
+);
